requirements.txt
flask==3.0.0
flask-socketio==5.3.6
eventlet==0.33.3
pyjwt==2.9.0
requests==2.32.3
gunicorn==21.2.0
